finetuned accelerometer

+++ MyApplication/app/src/main/java/com/shouryaaggarwal/hp/androidaccelerometerexample/Utils.java
@@ -10,13 +10,13 @@ public class Utils {
         void afterDelay();
     }
 
-    public static void delay(int secs, final DelayCallback delayCallback){
+    public static void delay(float secs, final DelayCallback delayCallback){
         Handler handler = new Handler();
         handler.postDelayed(new Runnable() {
             @Override
             public void run() {
                 delayCallback.afterDelay();
             }
-        }, secs * 1000); // afterDelay will be executed after (secs*1000) milliseconds.
+        }, (int)(secs * 1000)); // afterDelay will be executed after (secs*1000) milliseconds.
     }
 }
